+++ packages/libdxfrw-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libdxfrw-converter
 
+## 3.0.8
+
+### Patch Changes
+
+- feat: destory parser web worker once parsing finished
+- Updated dependencies
+  - @mlightcad/data-model@1.2.8
+
 ## 3.0.7
 
 ### Patch Changes