+++ packages/libredwg-converter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/libredwg-converter
 
+## 3.0.9
+
+### Patch Changes
+
+- feat: destory parser web worker once parsing finished
+- Updated dependencies
+  - @mlightcad/data-model@1.2.8
+
 ## 3.0.8
 
 ### Patch Changes